test: cover commander CLI behavior
