+++ test/unit/cli.test.ts
@@ -1,0 +1,89 @@
+import { describe, expect, test } from "bun:test";
+import { fileURLToPath } from "url";
+import { dirname, join } from "path";
+
+const entry = join(
+  dirname(fileURLToPath(import.meta.url)),
+  "..",
+  "..",
+  "src",
+  "index.ts",
+);
+
+async function runCli(args: string[]) {
+  const proc = Bun.spawn(["bun", "run", entry, ...args], {
+    stdout: "pipe",
+    stderr: "pipe",
+  });
+
+  const [stdout, stderr, exitCode] = await Promise.all([
+    new Response(proc.stdout).text(),
+    new Response(proc.stderr).text(),
+    proc.exited,
+  ]);
+
+  return {
+    stdout: stdout.trim(),
+    stderr: stderr.trim(),
+    exitCode,
+  };
+}
+
+describe("cli", () => {
+  test("hello prints the current text", async () => {
+    const result = await runCli(["hello"]);
+    expect(result.exitCode).toBe(0);
+    expect(result.stderr).toBe("");
+    expect(result.stdout).toBe("Hello, World!");
+  });
+
+  test("calc prints only the number result", async () => {
+    const result = await runCli(["calc", "2", "+", "3"]);
+    expect(result.exitCode).toBe(0);
+    expect(result.stderr).toBe("");
+    expect(result.stdout).toBe("5");
+  });
+
+  test("calc computes the modulo of two numbers", async () => {
+    const result = await runCli(["calc", "13", "%", "5"]);
+    expect(result.exitCode).toBe(0);
+    expect(result.stderr).toBe("");
+    expect(result.stdout).toBe("3");
+  });
+
+  test("help lists the available commands", async () => {
+    const result = await runCli(["--help"]);
+    expect(result.exitCode).toBe(0);
+    expect(result.stderr).toBe("");
+    expect(result.stdout).toContain(
+      "calc <left> <operator> <right>",
+    );
+    expect(result.stdout).toContain("hello");
+  });
+
+  test("calc rejects a non-numeric operand", async () => {
+    const result = await runCli(["calc", "two", "+", "3"]);
+    expect(result.exitCode).not.toBe(0);
+    expect(result.stdout).toBe("");
+    expect(result.stderr).toContain(
+      "error: command-argument value 'two' is invalid for argument 'left'",
+    );
+    expect(result.stderr).toContain("'two' is not a number.");
+  });
+
+  test("calc rejects an unsupported operator", async () => {
+    const result = await runCli(["calc", "2", "^", "3"]);
+    expect(result.exitCode).not.toBe(0);
+    expect(result.stdout).toBe("");
+    expect(result.stderr).toContain(
+      "Allowed choices are +, -, *, /, %",
+    );
+  });
+
+  test("rejects an unknown command", async () => {
+    const result = await runCli(["unknown"]);
+    expect(result.exitCode).not.toBe(0);
+    expect(result.stdout).toBe("");
+    expect(result.stderr).toContain("error: unknown command 'unknown'");
+  });
+});
